refactor self contained funtions
